LayoutTests/media/video-main-content-allow-then-scroll-expected.txt

RUN(internals.setMediaElementRestrictions(video, "RequireUserGestureForVideoRateChange,OverrideUserGestureRequirementForMainContent"))
EVENT(canplaythrough)
EVENT(playing)
RUN(document.scrollingElement.scrollTop = document.body.clientHeight)
EXPECTED (video.paused == 'false') OK
END OF TEST
